parikar/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
courlan==1.0.0
cron-descriptor==1.4.3
cryptography==42.0.2
cssselect==1.2.0
dateparser==1.2.0
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.0
django-celery-beat==2.5.0
django-contrib-comments==2.2.0
django-el-pagination==4.0.0
django-etc==1.4.0
django-hitcount==1.3.5
django-timezone-field==6.1.0
django-voting==1.1.0
gunicorn==21.2.0
htmldate==1.7.0
idna==3.6
image-pattern==0.0.18
joblib==1.3.2
jusText==3.0.0
kombu==5.3.5
langcodes==3.3.0
lxml==5.1.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
Pillow==7.2.0
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==1.10.14
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
readability-lxml==0.8.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
tld==0.13
trafilatura==1.7.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
